removed Subsystem factory
Working drivetrain code
