--- src/main/java/frc/robot/subsystems/SubsystemImpl.java
@@ -1,11 +1,0 @@
-package frc.robot.subsystems;
-
-import edu.wpi.first.wpilibj2.command.SubsystemBase;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface SubsystemImpl {
-    Class<? extends SubsystemBase> value();
-}
